Simplify the implementation of fp number to string cast
