+++ core/src/main/java/org/opensearch/sql/expression/function/PPLBuiltinOperators.java
@@ -77,8 +77,8 @@ import org.opensearch.sql.expression.function.udf.math.CRC32Function;
 import org.opensearch.sql.expression.function.udf.math.ConvFunction;
 import org.opensearch.sql.expression.function.udf.math.DivideFunction;
 import org.opensearch.sql.expression.function.udf.math.EulerFunction;
-import org.opensearch.sql.expression.function.udf.math.FormatNumberFunction;
 import org.opensearch.sql.expression.function.udf.math.ModFunction;
+import org.opensearch.sql.expression.function.udf.math.NumberToStringFunction;
 
 /** Defines functions and operators that are implemented only by PPL */
 public class PPLBuiltinOperators extends ReflectiveSqlOperatorTable {
@@ -356,7 +356,8 @@ public class PPLBuiltinOperators extends ReflectiveSqlOperatorTable {
       RELEVANCE_QUERY_FUNCTION_INSTANCE.toUDF("query_string", false);
   public static final SqlOperator MULTI_MATCH =
       RELEVANCE_QUERY_FUNCTION_INSTANCE.toUDF("multi_match", false);
-  public static final SqlOperator FORMAT_NUMBER = new FormatNumberFunction().toUDF("FORMAT_NUMBER");
+  public static final SqlOperator NUMBER_TO_STRING =
+      new NumberToStringFunction().toUDF("NUMBER_TO_STRING");
 
   /**
    * Returns the PPL specific operator table, creating it if necessary.
